Split v2 catalog publication module
